Switch to DeepseekV3ForCausalLM
